+++ api/src/main/kotlin/uy/kohesive/injekt/api/TypeInfo.kt
@@ -2,10 +2,6 @@ package uy.kohesive.injekt.api
 
 import java.lang.reflect.*
 
-// TODO: commented out code is because dealing with people wanting to key off of erased generic classes, is not so clear and clean.
-//       Kotlin MyClass::class.java does not create the same result as a class with generics does when reified.   Causes different
-//       results in the methods below.
-
 @Suppress("UNCHECKED_CAST")
 public fun Type.erasedType(): Class<Any> {
     return when (this) {
@@ -30,18 +26,12 @@ public fun Type.erasedType(): Class<Any> {
 
 public inline fun <reified T: Any> typeRef(): FullTypeReference<T> = object:FullTypeReference<T>(){}
 public inline fun <reified T: Any> fullType(): FullTypeReference<T> = object:FullTypeReference<T>(){}
-// public inline fun <reified T: Any> rawType(): TypeReference<T> = object:ErasedTypeReference<T>(){}
-// public inline fun <reified T: Any> eraseType(): TypeReference<T> = object:ErasedTypeReference<T>(){}
 
 public interface TypeReference<T> {
     public val type: Type
-//    public final val cyg: ClassWithGenerics
-//       get() = ClassWithGenerics.fromUnknown(type)
 }
 
 public abstract class FullTypeReference<T> protected constructor() : TypeReference<T> {
-    // public val erasedType: Class<Any> get() = type.eraseGenerics()
-
     override public val type: Type = javaClass.getGenericSuperclass() let { superClass ->
         if (superClass is Class<*>) {
             throw IllegalArgumentException("Internal error: TypeReference constructed without actual type information")
@@ -49,17 +39,6 @@ public abstract class FullTypeReference<T> protected constructor() : TypeReferen
         (superClass as ParameterizedType).getActualTypeArguments()[0]
     }
 }
-
-/*
-public abstract class ErasedTypeReference<T> protected constructor() : TypeReference<T> {
-    override public val type: Class<Any> = javaClass.getGenericSuperclass() let { superClass ->
-        if (superClass is Class<*>) {
-            throw IllegalArgumentException("Internal error: TypeReference constructed without actual type information")
-        }
-        (superClass as ParameterizedType).getActualTypeArguments()[0].eraseGenerics()
-    }
-}
-*/